A few days of progress
I started working on this project on the 5/24/2022 and I decided why not let anyone see this and have fun with it as well. I'll be streaming my progress in Youtube as Coder Cow so if you want to hop on in and chat feel free to join and let us code and learn together. :)